+++ Control/Joint.hs
@@ -1,7 +1,8 @@
 module Control.Joint (module Exports) where
 
-import Control.Joint.Operators as Exports
+import Control.Joint.Concepts as Exports
 import Control.Joint.Schemes as Exports
 import Control.Joint.Effects as Exports
 import Control.Joint.Abilities as Exports
+import Control.Joint.Operators as Exports
 import Control.Joint.Core as Exports