+++ Control/Joint/Concepts.hs
@@ -1,0 +1,3 @@
+module Control.Joint.Concepts (module Exports) where
+
+import Control.Joint.Concepts.Lens as Exports